fuzz: fix format handling

+++ spng.c
@@ -3363,7 +3363,7 @@ int spng_decoded_image_size(spng_ctx *ctx, int fmt, size_t *len)
     unsigned bytes_per_pixel;
 
     /* Currently all enums are single-bit values */
-    if(fmt & (fmt - 1)) return SPNG_EFMT; /* More than one bit is set */
+    if(fmt & ((unsigned)fmt - 1)) return SPNG_EFMT; /* More than one bit is set */
 
     if(fmt == SPNG_FMT_RGBA8)
     {
